+++ CHANGELOG.md
@@ -56,6 +56,7 @@ requirements (inherited from Drupal 11):
 - [Make the farmOS API optional #974](https://github.com/farmOS/farmOS/pull/973)
 - [Move QuantityCsvNormalizer to farm_csv module #977](https://github.com/farmOS/farmOS/pull/977)
 - [The farm_account_admin role has moved to a new farm_account_admin module](https://www.drupal.org/node/3527786)
+- [The farm_manager, farm_worker, and farm_viewer roles have moved to their own modules](https://www.drupal.org/node/3527787)
 
 ### Deprecated